+++ blockfrost-client/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Additions
   * `getAccountAssociatedAddressesTotal` for `/accounts/:stake_address/addresses/total`
   * `listPoolsExtended` for `/pools/extended`
+  * `/utils` API
+    * `deriveShelleyAddress` for `/utils/addresses/xpub/:xpub/:role/:index`
 
 # Version [0.7.1.1](https://github.com/blockfrost/blockfrost-haskell/compare/v0.7.1.0...client-0.7.1.1) (2023-01-10)